Add files via upload
Introduction
This system is basically concerned with the reservation is done.
In this we showed how the reservation is done.
What are the rules, regulation and procedure for reservation

Items Used :
Application Platform : Windows/Linux
Programming Language : C++
Libraries used : C++ Standard Library
Runs on : Console

In this project we are going to include : 
Bus numbers 
Class

Flow of System :
Reservation
Cancellation
Display status

Bugs :
On the time of writing this document, no bugs were discovered in the software. If you find any, please inform us about them. We would be very much happy to accept your fixes.

Conclusion :
At last we can say that the Bus Reservation system is the best process which is created a praiseworthy near our students and faculty's. In the time of our big occasion it can help us sit booking rapidly and arrival to our destination for our safe journey and create new era of modern science.